Create a workflow for downloading frames and extracting calibration envelopes from them.

Closes #15

See merge request asimov/pipelines/gwdata!21

+++ datafind/metafiles.py
@@ -2,11 +2,17 @@
 Functions to manipulate PESummary metafiles.
 """
 
+import shutil
+import logging
 import h5py
 import contextlib
 import subprocess
 import os
 import numpy as np
+
+from .calibration import CalibrationUncertaintyEnvelope
+
+logger = logging.getLogger("gwdata")
 
 class Metafile(contextlib.AbstractContextManager):
     """
@@ -25,9 +31,9 @@ class Metafile(contextlib.AbstractContextManager):
         """
         self.filename = filename
 
-    
+
     def __enter__(self):
-        
+
         self.metafile = h5py.File(self.filename)
 
         return self
@@ -47,6 +53,18 @@ class Metafile(contextlib.AbstractContextManager):
             psds[ifo] = PSD(psd, ifo=ifo)
         return psds
 
+    def calibration(self, analysis=None):
+        if not analysis:
+            # If no analysis is specified use the first one
+            analyses = list(self.metafile.keys())
+            analyses.remove("history")
+            analyses.remove("version")
+            analysis = sorted(analyses)[0]
+        cals = {}
+        for ifo, cal in self.metafile[analysis]['calibration_envelope'].items():
+            cals[ifo] = CalibrationUncertaintyEnvelope.from_array(cal)
+        return cals
+
 
 class PSD:
 
@@ -61,27 +79,26 @@ class PSD:
         tmp = "psd.tmp"
         self.to_ascii(tmp)
 
-        command = [
-            "convert_psd_ascii2xml",
-            "--fname-psd-ascii",
-            f"{tmp}",
-            "--conventional-postfix",
-            "--ifo",
-            f"{self.ifo}",
-        ]
+        executable = "convert_psd_ascii2xml"
+        if shutil.which(executable) is not None:
 
-        pipe = subprocess.Popen(
-            command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
-        )
-        out, err = pipe.communicate()
+            command = [
+                executable,
+                "--fname-psd-ascii",
+                f"{tmp}",
+                "--conventional-postfix",
+                "--ifo",
+                f"{self.ifo}",
+            ]
 
-        if err:
-            if hasattr(self.production.event, "issue_object"):
-                raise Exception(
-                    f"An XML format PSD could not be created.\n{command}\n{out}\n\n{err}",
-                )
-            else:
-                raise Exception(
-                    f"An XML format PSD could not be created.\n{command}\n{out}\n\n{err} ",
-                )
-        os.remove(tmp)
+            pipe = subprocess.Popen(
+                command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT
+            )
+            out, err = pipe.communicate()
+
+            if err:
+                logger.warning(f"An XML format PSD could not be created. {err}")
+            os.remove(tmp)
+
+        else:
+            logger.warning("An XML format PSD could not be created.")
